Fixed Issue #3 & removed idle players
Moved Invite button to separate template to stop selection of an
opponent re-rendering the selection box (this caused Issue #3). Later
when Meteor is updated to re-render based on element context this
workaround can be removed.
Fixed syntactical error in the remove_threshold check, so now idle
players will be removed after ten minutes.